# [feat] Show % of reference value in performance report