@victorusu 2021-07-01
@InternetUnexplorer, thanks for proposing this change! I like it! 
I just have a couple of questions. 
1. I am not sure how you want to deal with the `ref_val = 0`.
2. Not sure if the unit test will pass on every system.