@jjotero 2021-07-01
Hmm... I'm not sure there is a need for this. What's the benefit of reporting all this extra stuff? Usually, reference performance values are set on a ballpark approach, so knowing that a test performed at 90% or 120% relative to that value doesn't mean anything. And in those situations, all these extra data just buries the actual performance results with extra noise.

In fact, if you want that percentage reported, you can simply create a new performance variable that reports just that.