@brandongc 2021-07-01
@jjotero I don't think all reference values are so arbitrary. For example, something like stream has fairly definite value based on the hardware. Also, at NERSC we are working on filling our references out with more rigorous statistics now that we have a good collection of runs over time to draw from. 

As for the benefit, I think having the reference value available at the time of the report is helpful for the person running reframe to decide if the "ballpark reference" is reasonable, the percentage aspect is just a bonus for convenience.  Otherwise, currently I have to go look into the test and find what the reference was.