@InternetUnexplorer 2021-07-01
As @brandongc said, the goal is to make it easier to compare the values to the reference values, since currently one has to read the checks to see what the reference values are. At the same time, I definitely don't want to make the performance report significantly harder to read…