@vkarak 2021-07-02
Thanks @InternetUnexplorer for your PR and I think it's useful. I can see also @jjotero's point. His concern -if I interpret it correctly- is that often you might want to run a performance test on a new system as a benchmark, for which you don't have a reference and therefore showing anything related to reference values will seem like noise. But in this case, the reference value is set to zero (by convention), so the "reference" part of the output will not be printed (I will need to try it out). As of the formatting style, I'd vote for the "aligned everything."