@InternetUnexplorer 2021-07-02
Thanks for the feedback @vkarak and everyone else!

I made a few tweaks to the patch (still changing things a bit so haven't pushed yet) to make the formatting a sort of middle ground where things are partially aligned into three columns (the variables, the values, and the details). Here's what it looks like:
```text
PERFORMANCE REPORT
------------------------------------------------------------------------------
StreamWithRefTest
- prism:default
   - gnu
      * num_tasks: 1
      * Copy:  37339.6 MB/s (4% of 1000000 MB/s)
      * Scale: 24644.0 MB/s (49% of 50000 MB/s)
      * Add:   27235.6 MB/s (136% of 20000 MB/s)
      * Triad: 26993.7 MB/s (2699% of 1000 MB/s)
------------------------------------------------------------------------------
```
Note that all of the values here are the same length, but even if they weren't the details would still be aligned.

It probably wouldn't be too hard to align the different parts of the details as well, I just wasn't sure how useful it would actually be in practice since it already seems pretty easy to read in my opinion. However, I'm willing to give it a go if others want it.

Also, when making changes to this branch, should I amend and force-push, or just make new commits (and maybe squash them later?). I wasn't quite sure after reading the contribution guidelines.