@InternetUnexplorer 2021-07-07
Ok, I made the following changes:
- The name, value, and details sections are now vertically aligned
- The percentage value in the details section is left-padded to 3 digits
  - It's very simple this way, and I assume the probability of a value being ≥10x the reference value is quite low 
- If the reference value is present but zero, `N/A%` is displayed

The performance report now looks like this:
```text
PERFORMANCE REPORT
------------------------------------------------------------------------------
ExampleTest
- prism:default
   - gnu
      * num_tasks: 1
      * Unitless: 120.3         (120% of 100)
      * Foo:      27092.9 MB/s  (135% of 20000 MB/s)
      * ZeroRef:  1034.4 iter/s (N/A% of 0 iter/s)
------------------------------------------------------------------------------
```
The code currently checks whether the upper and lower bounds are present to tell the difference between the reference value being zero and the reference value being absent. I'm not sure if this is the correct way to do it, since it _is_ possible to pass `None` as both the upper and lower bounds, but I haven't been able to find a better way to do it.