@InternetUnexplorer 2021-07-07
Whoops, forgot to update `test_performance_report` to match, fixed.