@jjotero 2021-07-30
Hi @InternetUnexplorer ! Apologies for dragging this PR for so long. I wanted to say that we've been working on a new syntax to express the performance variables in regression tests, and this PR conflicts with some stuff we had in mind. Perhaps expanding a bit on my previous comment, some of the ReFrame users are just interested in logging performance data and not in comparing it with some reference. Currently, ReFrame enforces users to set a reference in order to get the performance variables from `perf_patterns` reported, and this is a limitation for those users just interested in the performance reporting part, which is what this new syntax that we have in the works is trying to address. This new syntax will make the `reference` field entirely optional, so I think it would be rather weird if we have this new field that you're proposing in the performance report which always expects the `reference` to be present. So if someone were to write a minimal ReFrame check without their reference, they would get that `(N/A% of 0<the_perf_var_units>)` by default.

I have a few ideas on how to integrate your proposed changes with the new syntax though, so I'd appreciate if you can have a look at #2083 and let us know what you think. This new perf syntax introduces the concept of a `performance_function` which is basically a decorated method which ReFrame automatically runs during the performance stage. So what I had in mind is that perhaps we could add an extra argument to this `performance_function` decorator where we pass a callable that produces whichever metadata the user wants into the performance report. This would make that additional field fully customisable on a per-variable basis. What do you think?