@brandongc 2021-07-30
Would that mean we would need to use this decorator syntax instead of the "simple"?

```python
        self.perf_patterns = {"flops":
                              sn.extractsingle(r'^(\S+\s){10}.*PERF_SUMMARY.*$', self.stdout, 1, float) }
```

Since most of our performance tests have references (or will) then we would need to make a callable, replicate/import that across all the tests and then pass it to each one?